backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Race management system for pinewood derby events"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
litestar = "^2.0.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.0"
sqlalchemy = "^2.0.0"
aiosqlite = "^0.19.0"
alembic = "^1.12.0"
pyjwt = "^2.8.0"
reportlab = "^4.0.4"
pillow = "^10.0.0"
pyserial-asyncio = "^0.6"
aiomysql = {version = "^0.2.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
ruff = "^0.0.287"

[tool.poetry.extras]
mysql = ["aiomysql"]

[tool.poetry.scripts]
start = "backend.main:run_app"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "B", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"